refaactor: move publisher to examples
